+++ lib/ex_json_schema/validator/const.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.Const do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"const", const}, data) do
     do_validate(const, data)
   end
