+++ lib/ex_json_schema/validator/contains.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.Contains do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(root, _, {"contains", contains}, data) do
     do_validate(root, contains, data)
   end
