+++ lib/ex_json_schema/validator/enum.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.Enum do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"enum", enum}, data) do
     do_validate(enum, data)
   end
