+++ lib/ex_json_schema/validator/exclusive_maximum.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.ExclusiveMaximum do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"exclusiveMaximum", maximum}, data) do
     do_validate(maximum, data)
   end
