+++ lib/ex_json_schema/validator/exclusive_minimum.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.ExclusiveMinimum do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"exclusiveMinimum", minimum}, data) do
     do_validate(minimum, data)
   end
