+++ lib/ex_json_schema/validator/format.ex
@@ -18,7 +18,6 @@ defmodule ExJsonSchema.Validator.Format do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"format", format}, data) do
     do_validate(format, data)
   end
