+++ lib/ex_json_schema/validator/items.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.Items do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(root, schema, {"items", _}, data) do
     do_validate(root, schema, data)
   end
