+++ lib/ex_json_schema/validator/max_length.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.MaxLength do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"maxLength", max_length}, data) do
     do_validate(max_length, data)
   end
