+++ lib/ex_json_schema/validator/max_properties.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.MaxProperties do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"maxProperties", max_properties}, data) do
     do_validate(max_properties, data)
   end
