+++ lib/ex_json_schema/validator/maximum.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.Maximum do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"maximum", maximum}, data) do
     do_validate(maximum, data)
   end
