+++ lib/ex_json_schema/validator/min_items.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.MinItems do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"minItems", min_items}, data) do
     do_validate(min_items, data)
   end
