+++ lib/ex_json_schema/validator/min_properties.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.MinProperties do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"minProperties", min_properties}, data) do
     do_validate(min_properties, data)
   end
