+++ lib/ex_json_schema/validator/multiple_of.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.MultipleOf do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"multipleOf", multiple_of}, data) do
     do_validate(multiple_of, data)
   end
