+++ lib/ex_json_schema/validator/not.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.Not do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(root, _, {"not", not_schema}, data) do
     do_validate(root, not_schema, data)
   end
