+++ lib/ex_json_schema/validator/one_of.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.OneOf do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(root, _, {"oneOf", one_of}, data) do
     do_validate(root, one_of, data)
   end
