+++ lib/ex_json_schema/validator/property_names.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.PropertyNames do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(root, _, {"propertyNames", property_names}, data) do
     do_validate(root, property_names, data)
   end
