+++ lib/ex_json_schema/validator/required.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.Required do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"required", required}, data) do
     do_validate(required, data)
   end
