+++ lib/ex_json_schema/validator/type.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.Type do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"type", type}, data) do
     do_validate(type, data)
   end
