+++ lib/ex_json_schema/validator/unique_items.ex
@@ -7,7 +7,6 @@ defmodule ExJsonSchema.Validator.UniqueItems do
 
   @impl ExJsonSchema.Validator
   @spec validate(Root.t(), ExJsonSchema.data(), {String.t(), ExJsonSchema.data()}, ExJsonSchema.data()) :: Validator.errors_with_list_paths
-
   def validate(_, _, {"uniqueItems", unique_items}, data) do
     do_validate(unique_items, data)
   end
