+++ test/json_schema_draft4_test_suite_test.exs
@@ -1,6 +1,6 @@
-# defmodule ExJsonSchema.JsonSchemaDraft4TestSuiteTest do
-#   use ExUnit.Case, async: true
+defmodule ExJsonSchema.JsonSchemaDraft4TestSuiteTest do
+  use ExUnit.Case, async: true
 
-#   use ExJsonSchema.Test.Support.TestSuiteTemplate,
-#     schema_tests_path: "test/JSON-Schema-Test-Suite/tests/draft4/"
-# end
+  use ExJsonSchema.Test.Support.TestSuiteTemplate,
+    schema_tests_path: "test/JSON-Schema-Test-Suite/tests/draft4/"
+end
