+++ test/support/test_suite_template.ex
@@ -1,17 +1,3 @@
-defmodule ExJsonSchema.JsonSchemaDraftTestSuiteTest.Helpers do
-
-  def schema_test_path(filename, schema_tests_path) do
-    Path.join(schema_tests_path, filename)
-  end
-
-  def load_schema_test(name, schema_tests_path) do
-    name <> ".json"
-    |> schema_test_path(schema_tests_path)
-    |> File.read!()
-    |> Poison.Parser.parse!()
-  end
-end
-
 defmodule ExJsonSchema.Test.Support.TestSuiteTemplate do
 
   use ExUnit.CaseTemplate
@@ -19,12 +5,12 @@ defmodule ExJsonSchema.Test.Support.TestSuiteTemplate do
   using opts do
     quote bind_quoted: [opts: opts] do
 
-      import ExJsonSchema.JsonSchemaDraftTestSuiteTest.Helpers
+      alias ExJsonSchema.Test.Support.TestHelpers
 
       @schema_tests_path opts[:schema_tests_path]
 
       @active [
-        # "dependencies with boolean subschemas: empty object is valid"
+        # "Recursive references between schemas: valid tree"
       ]
 
       "#{@schema_tests_path}**/*.json"
@@ -36,7 +22,7 @@ defmodule ExJsonSchema.Test.Support.TestSuiteTemplate do
           |> String.replace(".json", "")
 
         name
-        |> load_schema_test(@schema_tests_path)
+        |> TestHelpers.load_schema_test(@schema_tests_path)
         |> Enum.each(fn fixture ->
           %{"description" => description, "schema" => schema, "tests" => tests} = fixture
